abstract ModelStub to its own class
